OkTexHelper

Documentation for OkTexHelper.

OkTexHelper.blocklist

OkTexHelper.CleanComment

OkTexHelper.CleanComment

OkTexHelper.CleanComments

OkTexHelper.blocktable

OkTexHelper.cleanblock

OkTexHelper.cleancomment

OkTexHelper.isblock

OkTexHelper.matchbegin

OkTexHelper.matchend

OkTexHelper.repstar

OkTexHelper.writelines

OkTexHelper.writelines

OkTexHelper.blocklist — Constant

Tags of blocks to be detected in Tex (.tex) file.

OkTexHelper.CleanComment — Type

mutable struct CleanComment
    text::AbstractString
    pbreak::Bool
end

where pbreak denotes whether this line of text is originally line break in the document.

OkTexHelper.CleanComment — Method

CleanComment(str::AbstractString, tp) do cleancomment(str, tp), with whether this str is line break in the original text denoted in the pbreak field.

OkTexHelper.CleanComments — Method

Given the fpath to the text file, CleanComments(fpath) returns a Vector of CleanComment.

Example

CleanComments("temp/temp.tex") |> lines -> writelines(lines, "tempxxx.tex")

OkTexHelper.blocktable — Method

Given lines a vector of strings, blocktable(lines) returns vectors of isblock(lines, tag) for all tag in blocktag.

OkTexHelper.cleanblock — Method

Given file path fpath and column selector colselector, cleanblock(fpath, colselector) returns a vector of strings lines that blocks matching colselector were removed.

Use writelines to save your result as file.

Example

cleanblock("temp/temp.tex") |> lines -> writelines(lines, "tempxxx.tex"; keep_empty = false, do_rstrip=true)

OkTexHelper.cleancomment — Method

Example

cleancomment("temp/temp.tex") |> lines -> writelines(lines, "tempxxx.tex"; keep_empty = false, do_rstrip=true)

OkTexHelper.isblock — Method

Given lines a vector of strings, isblock(lines, tag::String) returns a Bool vector indicating the location of blocks matched by matchbegin and matchend.

OkTexHelper.matchbegin — Method

matchbegin(tag, str) returns true if str is started with \begin{$tag}

OkTexHelper.matchend — Method

matchend(tag, str) returns true if str is started with \end{$tag}

OkTexHelper.repstar — Method

repstar(str) add "\" in front of every asterisk.

OkTexHelper.writelines — Method

writelines(lines::Vector{String}, targetpath; keep_empty=true, do_rstrip=false) write a Vector of string to targetpath.

cleancomment may result in empty lines or lines with only white spaces, and sometimes these empty lines results in complie error of latex (e.g., an blank line in a table block). To avoid this, set keep_empty to false and do_rstrip to true to discard all empty lines.

OkTexHelper.writelines — Method

writelines(lines::Vector{CleanComment}, targetpath) write non-empty or non-line break lines of CleanComment to targetpath.

See also CleanComment, CleanComments.

Example

CleanComments("LATEX_gji2020/main.tex") |> lines -> writelines(lines, "temp/temp.tex")